+++ nonet/src/main/java/com/keiferstone/nonet/PollTask.java
@@ -15,22 +15,25 @@ import static com.keiferstone.nonet.ConnectionStatus.*;
 class PollTask extends AsyncTask<Void, Void, Integer> {
     private static final String TAG = PollTask.class.getSimpleName();
 
+    private static OkHttpClient client;
+
     private Configuration configuration;
     private OnPollCompletedListener listener;
-    private OkHttpClient client;
 
-    static void run(Configuration configuration, OnPollCompletedListener listener) {
+    static PollTask run(Configuration configuration, OnPollCompletedListener listener) {
         PollTask pollTask = new PollTask(configuration, listener);
-        pollTask.execute();
+        return (PollTask) pollTask.execute();
     }
 
     private PollTask(Configuration configuration, OnPollCompletedListener listener) {
         this.configuration = configuration;
         this.listener = listener;
 
-        client = new OkHttpClient.Builder()
-                .connectTimeout(configuration.getTimeout(), TimeUnit.SECONDS)
-                .build();
+        if (client == null) {
+            client = new OkHttpClient.Builder()
+                    .connectTimeout(configuration.getTimeout(), TimeUnit.SECONDS)
+                    .build();
+        }
     }
 
     @ConnectionStatus